feat: update base-component
