+++ src/client/components/component.ts
@@ -1,0 +1,7 @@
+export interface IComponent {
+  readonly element: HTMLElement;
+  beforeAppend?: () => Promise<void>;
+  onAppended?: () => Promise<void>;
+  beforeRemove?: () => Promise<void>;
+  onRemoved?: () => Promise<void>;
+}
