+++ src/client/root-component.ts
@@ -1,0 +1,3 @@
+export interface IRootComponent {
+  readonly rootElement: HTMLElement,
+}
